Update Maven compiler versions and simplify code
This commit updates the Maven compiler source and target versions from 21 to 22. It also simplifies the loop iterating through a list, switching from a for loop to using the forEach method. There are also corrections to the Rest Assured test by updating the expected IP address and some file name revisions for more accurate test identification. Lastly, a try-with-resource pattern is applied to handle IO operations with ZipFiles.